+++ src/services/auth.service.ts
@@ -1,31 +1,33 @@
-import { PrismaClient } from "@prisma/client";
-
+import prisma from "../../prisma/client";
 import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
 
-const prisma = new PrismaClient();
+const JWT_SECRET = process.env.JWT_SECRET || "supersecret";
+const JWT_EXPIRES_IN = "1h";
 
 export const registerUser = async (name: string, email: string, password: string) => {
-  const existingUser = await prisma.user.findUnique({
-    where: { email },
-  });
-
-  if (existingUser) {
-    throw new Error("El correo ya está registrado");
-  }
+  const existingUser = await prisma.user.findUnique({ where: { email } });
+  if (existingUser) throw new Error("El email ya está registrado");
 
   const hashedPassword = await bcrypt.hash(password, 10);
 
-  const newUser = await prisma.user.create({
-    data: {
-      name,
-      email,
-      password: hashedPassword,
-    },
+  const user = await prisma.user.create({
+    data: { name, email, password: hashedPassword },
   });
 
-  return {
-    id: newUser.id,
-    name: newUser.name,
-    email: newUser.email,
-  };
+  const token = jwt.sign({ userId: user.id }, JWT_SECRET, { expiresIn: JWT_EXPIRES_IN });
+
+  return { user: { id: user.id, name: user.name, email: user.email }, token };
+};
+
+export const loginUser = async (email: string, password: string) => {
+  const user = await prisma.user.findUnique({ where: { email } });
+  if (!user) throw new Error("Usuario no encontrado");
+
+  const isValid = await bcrypt.compare(password, user.password);
+  if (!isValid) throw new Error("Contraseña incorrecta");
+
+  const token = jwt.sign({ userId: user.id }, JWT_SECRET, { expiresIn: JWT_EXPIRES_IN });
+
+  return { user: { id: user.id, name: user.name, email: user.email }, token };
 };
